добавил скорость по тикам, как вы говорили, исправил все кнопки. Сделаны все пукнты с сайта

+++ kursovaya/kursovaya/Form1.cs
@@ -1,12 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
 using System.Drawing;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading;
 using System.Windows.Forms;
 
 namespace kursovaya
@@ -14,6 +9,8 @@ namespace kursovaya
     public partial class Form1 : Form
     {
         Emitter emitter;
+        bool ifRun = true;
+        bool stepPermission = false;
 
         public Form1()
         {
@@ -25,21 +22,23 @@ namespace kursovaya
                 width = picDisplay.Width,
                 gravitationY = 5
             };
-            timer1.Interval = 1000;
-        }     
+        }
 
         private void timer1_Tick(object sender, EventArgs e)
         {
-            emitter.updateState();
+            if ((speedBar.Value != 0 && ifRun) || stepPermission)
+            {
+                emitter.updateState();
+            }
             using (var g = Graphics.FromImage(picDisplay.Image))
             {
                 g.Clear(Color.Black);
                 emitter.render(g);
             }
             picDisplay.Invalidate();
-            label2.Text = emitter.particlesHistory[emitter.currentHistoryIndex - 1][0].ToString();
+            stepPermission = false;      
         }
-        
+
         private void picDisplay_MouseMove(object sender, MouseEventArgs e)
         {
             emitter.X = e.X;
@@ -50,15 +49,15 @@ namespace kursovaya
             if (emitter.ifInCircle(out circleX, out circleY, out circleRadius, out life))
             {
                 Graphics circle = picDisplay.CreateGraphics();
-                drawCircle(circle, circleX-circleRadius, circleY-circleRadius, circleRadius);
+                drawCircle(circle, circleX - circleRadius, circleY - circleRadius, circleRadius);
                 showCircleInfo(circle, circleX, circleY, circleRadius, life);
             }
         }
-        
+
         private void drawCircle(Graphics g, float x, float y, int radius)
         {
             Pen pen = new Pen(Brushes.Red);
-            g.DrawEllipse(pen, x, y, radius*2, radius*2);
+            g.DrawEllipse(pen, x, y, radius * 2, radius * 2);
         }
 
         private void showCircleInfo(Graphics g, float x, float y, int radius, float life)
@@ -76,10 +75,48 @@ namespace kursovaya
 
         private void speedBar_ValueChanged(object sender, EventArgs e)
         {
-            if (speedBar.Value == 0) { timer1.Enabled = false; return; }
-            timer1.Enabled = true;
-            timer1.Interval = 1000 - 100 * speedBar.Value;
-            label1.Text = timer1.Interval.ToString();
+            ifRun = true;
+            setTickRate();
+        }
+
+        private void setTickRate()
+        {
+            switch (speedBar.Value)
+            {
+                case 0:
+                    ifRun = false;
+                    break;
+                case 1:
+                    emitter.tickRate = 30;
+                    break;
+                case 2:
+                    emitter.tickRate = 25;
+                    break;
+                case 3:
+                    emitter.tickRate = 20;
+                    break;
+                case 4:
+                    emitter.tickRate = 15;
+                    break;
+                case 5:
+                    emitter.tickRate = 10;
+                    break;
+                case 6:
+                    emitter.tickRate = 7;
+                    break;
+                case 7:
+                    emitter.tickRate = 5;
+                    break;
+                case 8:
+                    emitter.tickRate = 3;
+                    break;
+                case 9:
+                    emitter.tickRate = 2;
+                    break;
+                case 10:
+                    emitter.tickRate = 1;
+                    break;
+            }
         }
 
         public void drawSpeedVector()
@@ -91,55 +128,57 @@ namespace kursovaya
                 int deviation = (int)(particle.speedX * 9);
                 Pen pen = new Pen(Brushes.Green);
                 speedVector.DrawLine(pen, new Point((int)particle.x, (int)particle.y),
-                    new Point((int)(particle.x + particle.radius * Math.Cos(deviation - 90)), 
+                    new Point((int)(particle.x + particle.radius * Math.Cos(deviation - 90)),
                     (int)(particle.y + particle.radius * Math.Sin(deviation - 90))));
             }
         }
 
         private void stopButton_Click(object sender, EventArgs e)
         {
-            timer1.Stop();
+            ifRun = false;
         }
 
         private void stepButton_Click(object sender, EventArgs e)
         {
-            if (emitter.currentHistoryIndex < emitter.particlesHistory.Count)
+            ifRun = false;
+            if (emitter.currentHistoryIndex < emitter.particlesHistory.Count-1 && emitter.currentHistoryIndex != 19)
             {
                 //поставить значения дальше по списку
+                emitter.particles.RemoveRange(0, emitter.particles.Count);
+                foreach (ParticleColorful particle in emitter.particlesHistory[emitter.currentHistoryIndex + 1])
+                {
+                    ParticleColorful part = new ParticleColorful(particle);
+                    emitter.particles.Add(part);
+                }
                 emitter.currentHistoryIndex++;
             }
             else
             {
-                timer1.Stop();
-                timer1_Tick(sender, e);
-                timer1.Stop();
+                emitter.tickCount += (emitter.tickRate - emitter.tickCount % emitter.tickRate);
+                stepPermission = true;
             }
         }
 
+        //СТАРТ
         private void startButton_Click(object sender, EventArgs e)
         {
-            if (emitter.currentHistoryIndex != 19)
-            {
-                //поставить значения дальше по списку
-                emitter.currentHistoryIndex++;
-            }
-            else timer1.Start();
+            ifRun = true;
+            setTickRate();
         }
 
         private void backButton_Click(object sender, EventArgs e)
         {
-            if (emitter.currentHistoryIndex >= 1)
+            ifRun = false;
+            //backProhibition = true;
+            if (emitter.currentHistoryIndex >= 2)
             {
                 //вернуться на значения из списка
-                emitter.particles.RemoveRange(0, emitter.particles.Count);              
-                foreach (Particle particle in emitter.particlesHistory[emitter.currentHistoryIndex-1])
+                emitter.particles.RemoveRange(0, emitter.particles.Count);
+                foreach (ParticleColorful particle in emitter.particlesHistory[emitter.currentHistoryIndex - 2])
                 {
-                    Particle part = new Particle(particle);
+                    ParticleColorful part = new ParticleColorful(particle);
                     emitter.particles.Add(part);
                 }
-                
-                var g = Graphics.FromImage(picDisplay.Image);
-                emitter.render(g);
                 emitter.currentHistoryIndex--;
             }
         }
@@ -147,10 +186,10 @@ namespace kursovaya
         private void listButton_Click(object sender, EventArgs e)
         {
             using (StreamWriter sw = new StreamWriter(@"D:\test.txt", false, System.Text.Encoding.Default))
-            { 
+            {
                 string text = "";
-                
-                foreach (List<Particle> list in emitter.particlesHistory)
+
+                foreach (List<ParticleColorful> list in emitter.particlesHistory)
                 {
                     foreach (Particle particle in list)
                     {
